Create popup to add banned pages
Adds a browser action popup that allows user to add new pages to
doomscrolling list.

Adds the hostname of the page to the list which makes it so the user can
use the button on a subdirectory and the extension bans the whole site.
With this user can't ban subdirectories individually (i.e. can't ban
only youtube.com/shorts/ instead of whole of youtube.com) but subdomains
work independently (banning fi.wikipedia.org doesn't ban en.wikipedia.org).